+++ crates/components-runtime/src/context.rs
@@ -6,9 +6,10 @@ use wasmtime::{
 };
 use wasmtime_wasi::{IoView, ResourceTable, WasiCtx, WasiView};
 
-use crate::config::{ComponentsConfiguration, DataCollectionComponents};
+use crate::config::ComponentsConfiguration;
 use crate::consent_mapping::{ConsentMapping, ConsentMappingPre};
-use crate::data_collection::v1_0_0::data_collection::{DataCollection, DataCollectionPre};
+use crate::data_collection::v1_0_0::data_collection::DataCollectionPre;
+use crate::data_collection::v1_0_0::pre_instanciate_data_collection_component_1_0_0_internal;
 use crate::data_collection::version::DataCollectionWitVersion;
 
 pub struct ComponentsContext {
@@ -19,27 +20,6 @@ pub struct ComponentsContext {
 pub struct Components {
     pub data_collection_1_0_0: HashMap<String, DataCollectionPre<HostState>>,
     pub consent_mapping: HashMap<String, ConsentMappingPre<HostState>>,
-}
-
-pub fn pre_instanciate_data_collection_component_1_0_0_internal(
-    engine: &Engine,
-    component_config: &DataCollectionComponents,
-) -> anyhow::Result<DataCollectionPre<HostState>> {
-    let mut linker = Linker::new(engine);
-    wasmtime_wasi::add_to_linker_async(&mut linker)?;
-
-    let span = tracing::info_span!("component-context", component = %component_config.id, category = "data-collection");
-    let _span = span.enter();
-
-    tracing::debug!("Loading new data collection component");
-
-    let component = Component::from_file(engine, &component_config.file)?;
-    let instance_pre = linker.instantiate_pre(&component)?;
-    let instance_pre = DataCollectionPre::new(instance_pre)?;
-
-    tracing::debug!("loaded new data collection component");
-
-    Ok(instance_pre)
 }
 
 impl ComponentsContext {
@@ -99,53 +79,12 @@ impl ComponentsContext {
         Ok(Self { engine, components })
     }
 
-    pub fn pre_instanciate_data_collection_1_0_0_component(
-        &self,
-        component_config: DataCollectionComponents,
-    ) -> anyhow::Result<DataCollectionPre<HostState>> {
-        let instance_pre = pre_instanciate_data_collection_component_1_0_0_internal(
-            &self.engine,
-            &component_config,
-        )?;
-        Ok(instance_pre)
-    }
-
-    pub fn add_data_collection_1_0_0_component(
-        &mut self,
-        component_config: DataCollectionComponents,
-        instance_pre: DataCollectionPre<HostState>,
-    ) {
-        if !self
-            .components
-            .data_collection_1_0_0
-            .contains_key(&component_config.id)
-        {
-            self.components
-                .data_collection_1_0_0
-                .insert(component_config.id.clone(), instance_pre);
-        }
-    }
-
     pub fn empty_store(&self) -> Store<HostState> {
         Store::new(&self.engine, HostState::new())
     }
 
     pub fn empty_store_with_stdout(&self) -> Store<HostState> {
         Store::new(&self.engine, HostState::new_with_stdout())
-    }
-
-    pub async fn get_data_collection_1_0_0_instance(
-        &self,
-        id: &str,
-        store: &mut Store<HostState>,
-    ) -> anyhow::Result<DataCollection> {
-        let instance_pre = self.components.data_collection_1_0_0.get(id);
-
-        if instance_pre.is_none() {
-            return Err(anyhow::anyhow!("component not found: {}", id));
-        }
-
-        instance_pre.unwrap().instantiate_async(store).await
     }
 
     pub async fn get_consent_mapping_instance(
